content/base/test/test_bug840098.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=840098
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 840098</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=840098">Mozilla Bug 840098</a>
<p id="display"></p>
<div id="content" style="display: none">
  <div id="foo"></div>
</div>
<marquee id="m">Hello</marquee>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 840098 **/

var mar = document.getElementById("m");
var anonymousNode = SpecialPowers.unwrap(SpecialPowers.wrap(document).getAnonymousNodes(mar)[0]);
try {
  document.implementation.createDocument("", "", null).adoptNode(anonymousNode);
  ok(false, "shouldn't be able to adopt the root of an anonymous subtree");
} catch (e) {
  is(e.name, "NotSupportedError", "threw the correct type of error");
}

</script>
</pre>
</body>
</html>
